+++ communication/src/one/microstream/com/ComProtocolData.java
@@ -24,6 +24,6 @@ public interface ComProtocolData extends PersistenceTypeDictionaryViewProvider
 		return this.typeDictionary();
 	}
 
-	public long inactivityTimeout();
+	public int inactivityTimeout();
 	
 }
